refactor: UserController에 사용자 목록 조회 및 프로필 조회 기능 통합
- UserListController와 UserProfileController를 UserController에 통합
