--- src/main/java/codesquad/springcafe/web/controller/UserListController.java
@@ -1,26 +1,0 @@
-package codesquad.springcafe.web.controller;
-
-import codesquad.springcafe.web.domain.User;
-import codesquad.springcafe.web.service.UserService;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-import java.util.List;
-
-@Controller
-public class UserListController {
-
-    private final UserService userService;
-
-    public UserListController(UserService userService) {
-        this.userService = userService;
-    }
-
-    @GetMapping("/user/list")
-    public String userList(Model model) {
-        List<User> users = userService.findUsers();
-        model.addAttribute("users", users);
-        return "user/list";
-    }
-}
